Extract concern for state_machine helpers
So can access these helper methods in other models we have with state
machines, and to give a place to add future shared behaviour (which will
soon be useful).
