+++ app/models/change_motd_request.rb
@@ -1,4 +1,6 @@
 class ChangeMotdRequest < ApplicationRecord
+  include HasStateMachine
+
   validates_presence_of :motd, :state
   belongs_to :case
