+++ app/models/concerns/has_state_machine.rb
@@ -1,0 +1,16 @@
+
+module HasStateMachine
+  extend ActiveSupport::Concern
+
+  included do
+    class << self
+      def events
+        state_machine.events.keys
+      end
+
+      def possible_states
+        state_machine.states.keys
+      end
+    end
+  end
+end
