+++ app/models/maintenance_window.rb
@@ -1,4 +1,6 @@
 class MaintenanceWindow < ApplicationRecord
+  include HasStateMachine
+
   default_scope { order(created_at: :desc) }
   belongs_to :case
   belongs_to :cluster, required: false
@@ -48,14 +50,6 @@ class MaintenanceWindow < ApplicationRecord
   end
 
   class << self
-    def events
-      state_machine.events.keys
-    end
-
-    def possible_states
-      state_machine.states.keys
-    end
-
     # A maintenance window is 'finished' once it has reached a state which it
     # cannot transition out of.
     def finished_states
